fix(ci): unbreak CI on Python 3.10 + mypy strict
Four fixes prompted by the post-P2 CI failures:

* rule_deck.py — Python 3.10 has no stdlib 'tomllib'. Use a version-
  guarded import: tomllib (3.11+) or tomli (3.10). Add tomli as a
  3.10-only dependency ("tomli>=2.0; python_version<'3.11'"). The
  test_load_toml test now passes on the 3.10 CI matrix.
* gauges.py — explicit float() cast on weighted_rms_epe return so mypy
  no longer sees Any leak through Python's exponent overload.
* optics.py — type _pil_to_float32 as np.ndarray (under TYPE_CHECKING)
  so the .copy() call type-checks. Was previously typed as 'object'.
* export_cmd.py — torch.jit.trace is untyped in torch's stubs; add a
  scoped 'no-untyped-call' ignore (we can't fix it upstream from here).

No behavior change. Verified: ruff check + format --check + mypy strict
+ all 526 tests pass locally on 3.12.

+++ src/openlithohub/workflow/gauges.py
@@ -99,7 +99,7 @@ class GaugeTable:
         if wsum == 0.0:
             raise ValueError("All gauge weights are zero; weighted RMS is undefined.")
         num = sum(p.weight * e * e for p, e in zip(self.points, epes, strict=True))
-        return (num / wsum) ** 0.5
+        return float((num / wsum) ** 0.5)
 
 
 def parse_gauge(path: str | Path) -> GaugeTable:
